add contacts for ask question

+++ handlers/state/question.py
@@ -4,16 +4,18 @@ from aiogram.filters.command import Command
 from aiogram.fsm.context import FSMContext
 from aiogram.fsm.state import State, StatesGroup
 from aiogram.types import CallbackQuery, Message, ReplyKeyboardRemove
+from aiogram.utils.markdown import hlink
 from asgiref.sync import sync_to_async
 
-from common import send_admin_notification
-from config import Action, MenuAction
+from common import aupdate_user_info, send_admin_notification
+from config import Action, MenuAction, MenuKeyboard
 from ORM.models import Question as UserQuestion, User
 from run import bot, dp
 
 
 class Question(StatesGroup):
     question = State()
+    phone_number = State()
 
 def user_authorized(function):
     async def wrapper(callback_query, *args, **kwargs):
@@ -27,6 +29,20 @@ def user_authorized(function):
     return wrapper
 
 
+@dp.message(Question.question and Command(commands=['cancel']))
+async def cansel(message: Message, state: FSMContext) -> None:
+    data = await state.get_data()
+    info_msg_id = data.get('info_msg_id')
+    if info_msg_id:
+        await bot.delete_message(chat_id=message.chat.id, message_id=info_msg_id)
+
+    try:
+        await message.delete()
+    except TelegramBadRequest:
+        pass
+
+    await state.clear()
+    await message.answer('Скасовано',reply_markup=ReplyKeyboardRemove(),)
 
 @dp.callback_query(MenuAction.filter(F.action == Action.ask_question.value))
 @user_authorized
@@ -50,46 +66,48 @@ async def ask(callback_query: CallbackQuery, callback_data: MenuAction, state: F
     await callback_query.answer()
 
 
-@dp.message(Question.question and Command(commands=['cancel']))
-async def cansel(message: Message, state: FSMContext) -> None:
-    data = await state.get_data()
-    info_msg_id = data.get('info_msg_id')
-    if info_msg_id:
-        await bot.delete_message(chat_id=message.chat.id, message_id=info_msg_id)
 
-    try:
-        await message.delete()
-    except TelegramBadRequest:
-        pass
-
-    await state.clear()
-    await message.answer('Скасовано',reply_markup=ReplyKeyboardRemove(),)
 
 @dp.message(Question.question)
 async def process_name(message: Message, state: FSMContext) -> None:
+    await message.answer('Поділіться своїм контактом, якщо підтверджуєте надсилання запитання\n\nДля скасування: /cancel',
+                         reply_markup=MenuKeyboard().get_share_contact())
+    await state.update_data(init_message=message.text)
+    await state.set_state(Question.phone_number)
 
-    notification_message = [
-        f'#питання',
-        message.text,
-        f'',
-        f'Користувач: {message.chat.full_name} ',
-        f'@{message.chat.username} (#{message.chat.id})',
-    ]
 
-    sent_messages = await send_admin_notification('\n'.join(notification_message))
+@dp.message(Question.phone_number)
+async def phone_number(message: Message, state: FSMContext) -> None:
 
-    if sent_messages:
-        users_data = await sync_to_async(list)(User.objects.filter(id=message.from_user.id))
-        if users_data:
-            user = users_data[0]
-            q = await UserQuestion.objects.acreate(text=message.text, user=user, type='question')
-        await message.answer('Ваше запитання було надіслано')
+    if message.contact:
+        user, is_created = await aupdate_user_info(message)
+        data = await state.get_data()
+
+        clean_phone_number = '+' + message.contact.phone_number.replace("+", "")
+        notification_message = [
+            f'#питання',
+            data.get('init_message', '-'),
+            f'',
+            f'Користувач: {message.chat.full_name} ',
+            f'@{message.chat.username} (#{message.chat.id})',
+            f'''Телефон: {hlink(clean_phone_number, url=f'http://t.me/{clean_phone_number}')}''',
+        ]
+
+        sent_messages = await send_admin_notification('\n'.join(notification_message))
+
+        if sent_messages:
+            q = await UserQuestion.objects.acreate(text=data.get('init_message', '-'), user=user, type='question')
+            await message.answer('Ваше запитання було надіслано', reply_markup=ReplyKeyboardRemove())
+        else:
+            await message.answer('Ваше запитання не було надіслано', reply_markup=ReplyKeyboardRemove())
+
+        data = await state.get_data()
+        info_msg_id = data.get('info_msg_id')
+        if info_msg_id:
+            await bot.delete_message(chat_id=message.chat.id, message_id=info_msg_id)
+
+        await state.clear()
     else:
-        await message.answer('Ваше запитання не було надіслано')
+        await message.answer('Ви маєте поділитися номером телефону саме за кнопкою нижче, якщо підтверджуєте замовлянння\n\nДля скасування: /cancel',
+                             reply_markup=MenuKeyboard().get_share_contact())
 
-    data = await state.get_data()
-    info_msg_id = data.get('info_msg_id')
-    if info_msg_id:
-        await bot.delete_message(chat_id=message.chat.id, message_id=info_msg_id)
-
-    await state.clear()
